ml/requirements.txt
lightgbm==4.3.0
xgboost==2.0.3
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0
shap==0.45.0
prophet==1.1.5
optuna==3.5.0
mlflow==2.10.0
google-cloud-bigquery==3.25.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.42.0
pyarrow==15.0.0
